feat: text utils file
